API fix, refactor, cleanup
