+++ src/app/Triggers/CaptureSmartTriggers.tsx
@@ -14,24 +14,37 @@
  * limitations under the License.
  */
 import { BreadcrumbPage } from '@app/BreadcrumbPage/BreadcrumbPage';
+import { NullableTarget } from '@app/Shared/Services/api.types';
+import { ServiceContext } from '@app/Shared/Services/Services';
 import { TargetContextSelector } from '@app/TargetView/TargetContextSelector';
+import { useSubscriptions } from '@app/utils/hooks/useSubscriptions';
 import { Card, CardBody, Stack, StackItem } from '@patternfly/react-core';
 import * as React from 'react';
-import { TriggersTable } from './Triggers';
+import { of } from 'rxjs';
+import { SmartTriggersTable } from './SmartTriggers';
 
 export interface CaptureTriggersProps {}
 
 export const CaptureSmartTriggers: React.FC<CaptureTriggersProps> = ({ ...props }) => {
+  const context = React.useContext(ServiceContext);
+  const addSubscription = useSubscriptions();
+  const [target, setTarget] = React.useState(undefined as NullableTarget);
+  const targetAsObs = React.useMemo(() => of(target), [target]);
+
+  React.useEffect(() => {
+    addSubscription(context.target.target().subscribe((t) => setTarget(t)));
+  }, [addSubscription, context.target, setTarget]);
+
   const cardBody = React.useMemo(
     () => (
       <Stack hasGutter>
         <StackItem>
           <TargetContextSelector />
         </StackItem>
-        <StackItem>{<TriggersTable />}</StackItem>
+        <StackItem>{<SmartTriggersTable target={targetAsObs} isNestedTable={false} />}</StackItem>
       </Stack>
     ),
-    [],
+    [targetAsObs],
   );
 
   return (
